// src/pages/pageHome/components/desktop/Desktop.types.tsx
export type IAppId =
	| "settings"
	| "user"
	| "calculator"
	| "camera"
	| "clock"
	| "notes"
	| "tetris"
	| "snake"
	| "clashRoyale"
	| "test"
	| "testSelect"
	| "testTree"
	| "testTabs"
	| "testTable"
	| "testTransition";

export type IRole = "guest" | "user" | "admin";
export type IAuth = "loggedIn" | "loggedOut";